fix compare bug

+++ debug/accuracy_tools/api_accuracy_checker/bench_functions/rms_norm.py
@@ -6,7 +6,7 @@ from api_accuracy_checker.common.function_factory import npu_custom_functions, n
 def npu_rms_norm(x, gamma, epsilon=1e-5):
     rstd = torch.rsqrt(torch.mean(torch.pow(x, 2), axis=-1, keepdim=True) + epsilon)
     res = x * rstd * gamma
-    return res.cpu(), rstd.cpu()
+    return res.cpu(), rstd.float().cpu()
 
 
 @npu_custom_grad_functions
